refactor: move data set dimensions stuff to data set action/reducer
