+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-12T15:06:15.331Z\n"
-"PO-Revision-Date: 2019-02-12T15:06:15.331Z\n"
+"POT-Creation-Date: 2019-02-12T15:30:21.040Z\n"
+"PO-Revision-Date: 2019-02-12T15:30:21.040Z\n"
 
 msgid "Cancel"
 msgstr ""
